117 ensure fresh install works even if ifadv missing (#121)
* Remove ifadv from dependencies

It is not an official CRAN dependency

* Hot fix the problem with ifadv

* Hot fix the problem with ifadv without using devtools

* Improvements in get_ifadv

But still causing odd problems: tests go fine on testing, but fail on checking

* Greatly simplify the download process

For some reason, native R data files (.rda) are less comfortable for R than the good-old csv. I pushed a csv copy of the rda file to https://github.com/elpaco-escience/ifadv/tree/csv, and that's the one we are using now.